refactor: extract duplicated importNamePairs into shared import-utils.ts (#1826)
docs check acknowledged — internal dedup of an existing helper, no new
functionality, public API, or architecture change.

Impact: 1 functions changed, 22 affected
