+++ src/domain/graph/builder/incremental.ts
@@ -16,7 +16,6 @@ import type {
   BetterSqlite3Database,
   EngineOpts,
   ExtractorOutput,
-  Import,
   PathAliases,
   SqliteStatement,
 } from '../../../types.js';
@@ -30,6 +29,7 @@ import {
   resolveSameClassQualifiedMethod,
 } from './call-resolver.js';
 import { BUILTIN_RECEIVERS, readFileSafe } from './helpers.js';
+import { importNamePairs } from './import-utils.js';
 
 // ── Local types ─────────────────────────────────────────────────────────
 
@@ -317,23 +317,6 @@ function resolveBarrelTarget(
     }
   }
   return null;
-}
-
-/**
- * Pairs each locally-bound name from an import statement with its original
- * (pre-rename) exported name — identical to the local name unless the
- * specifier renames a binding (`import { X as Y }`). Barrel tracing and
- * target-file symbol lookups must search using the *original* name — the
- * renamed local alias only exists in the importing file, not in the file
- * being imported from (#1730). Mirrors `importNamePairs` in build-edges.ts.
- */
-function importNamePairs(imp: Import): Array<{ local: string; original: string }> {
-  const originalNameFor = new Map<string, string>();
-  for (const r of imp.renamedImports ?? []) originalNameFor.set(r.local, r.imported);
-  return imp.names.map((name) => {
-    const local = name.replace(/^\*\s+as\s+/, '');
-    return { local, original: originalNameFor.get(local) ?? local };
-  });
 }
 
 /**
