+++ src/domain/graph/builder/stages/build-edges.ts
@@ -56,6 +56,7 @@ import {
   CHA_TYPED_DISPATCH_CONFIDENCE,
   runChaPostPass,
 } from '../helpers.js';
+import { importNamePairs } from '../import-utils.js';
 import { getResolved, isBarrelFile, resolveBarrelExportCached } from './resolve-imports.js';
 
 // ── Local types ──────────────────────────────────────────────────────────
@@ -146,23 +147,6 @@ function importEdgeKind(imp: Import): string {
   if (imp.typeOnly) return 'imports-type';
   if (imp.dynamicImport) return 'dynamic-imports';
   return 'imports';
-}
-
-/**
- * Pairs each locally-bound name from an import statement with its original
- * (pre-rename) exported name — identical to the local name unless the
- * specifier renames a binding (`import { X as Y }`). Barrel tracing and
- * target-file symbol lookups must search using the *original* name — the
- * renamed local alias only exists in the importing file, not in the file
- * being imported from (#1730).
- */
-function importNamePairs(imp: Import): Array<{ local: string; original: string }> {
-  const originalNameFor = new Map<string, string>();
-  for (const r of imp.renamedImports ?? []) originalNameFor.set(r.local, r.imported);
-  return imp.names.map((name) => {
-    const local = name.replace(/^\*\s+as\s+/, '');
-    return { local, original: originalNameFor.get(local) ?? local };
-  });
 }
 
 /**
